+++ MapView/Map/RMMapViewDelegate.h
@@ -37,8 +37,8 @@
 - (void) beforeMapMove: (RMMapView*) map;
 - (void) afterMapMove: (RMMapView*) map ;
 
-- (void) beforeMapZoom: (RMMapView*) map byFactor: (float) zoomFactor near:(CGPoint) center;
-- (void) afterMapZoom: (RMMapView*) map byFactor: (float) zoomFactor near:(CGPoint) center;
+- (void) beforeMapZoom: (RMMapView*) map byFactor: (double)zoomFactor near:(CGPoint) center;
+- (void) afterMapZoom: (RMMapView*) map byFactor: (double) zoomFactor near:(CGPoint) center;
 
 - (void) doubleTapOnMap: (RMMapView*) map At: (CGPoint) point;
 - (void) singleTapOnMap: (RMMapView*) map At: (CGPoint) point;
