+++ MapView/Map/RMMarker.m
@@ -198,7 +198,7 @@ static CGImageRef _markerBlue = nil;
 	[super dealloc];
 }
 
-- (void)zoomByFactor: (float) zoomFactor near:(CGPoint) center
+- (void)zoomByFactor: (double) zoomFactor near:(CGPoint) center
 {
 	self.position = RMScaleCGPointAboutPoint(self.position, zoomFactor, center);
